Refactor and update global code structure
