+++ legacy_files/wallet_ver2.py
@@ -1,10 +1,10 @@
 import re
-import datetime
+from datetime import datetime, timedelta
 import random
 from abc import ABC
+import sqlite3
 
-"""Version 1.0"""
-"""Wallet without DB"""
+"""db python structures"""
 
 """
                                                 Database TABLE
@@ -27,47 +27,75 @@ From_number: Sender's account number (integer).
 To_number: Receiver's account number (integer).
 TransactionDate: Transaction date (in the format "yyyy.mm.dd").
 TransactionID: Unique transaction identifier (text).
-Amount: Transaction amount (float)."""
+Amount: Transaction amount (float).
 
-# list of info about accounts
-user_accounts = {
-    "12345678": {"type": "Дебетовий", "name": "John Smith", "transactions": [], "balance": 0.0},
-    "87654321": {"type": "Кредитний", "name": "Jane Doe", "transactions": [], "balance": 0.0},
-    "65432198": {"type": "Дебетовий", "name": "Michael Johnson", "transactions": [], "balance": 0.0}
-}
+Category Table
+Name: category name
+"""
+
+"""Сделать API для коннекта с db"""
+
+
+class DatabaseManager:
+    def __init__(self, database_path):
+        self.conn = None
+        self.database_path = database_path
+
+    def connect(self):
+        try:
+            self.conn = sqlite3.connect(self.database_path)
+            if self.conn:
+                print('Підключення до бази даних пройшло успішно!')
+
+        except sqlite3.Error as e:
+            print('Виникла помилка під час підключення до бази даних:', str(e))
+
+    def create_cursor(self):
+        if self.conn:
+            print(f"Курсор бази данних для шляха {self.database_path} створено")
+            return self.conn.cursor()
+
+    def commit(self):
+        self.conn.commit()
+
+    def close(self):
+        print("db Close")
+        if self.conn:
+            self.conn.close()
+
+
+# ----------------------------------------PATH DATABASE FILE----------------------------------------------------
+# database_path = input("Database path: ")
+db = DatabaseManager("/db\\Wallet.db")
+db.connect()
+cursor = db.create_cursor()
+
+# --------------------------------------------------------------------------------------------------------------
+cursor.execute("SELECT Name FROM Category")
+categories = cursor.fetchall()
+
+# lst standard categories and user categories
+user_categories = [category[0] for category in categories]
+
+# ---------------------------------------------------------------------------------------------------------------
+cursor.execute("SELECT Number FROM User_Accounts")
+numbers = cursor.fetchall()
+
 # lst number accounts for verification
-lst_accounts = ["12345678", "87654321", "65432198"]
-
-# lst name categories for verification
-# оставить список и потом из базы данных наполнять список недостающими
-user_categories = ["Їжа", "Одяг і взуття", "Розваги", "Транспорт", "Комунальні послуги", "Оплата житла",
-                   "Здоров'я та медицина", "Подарунки та благодійність", "Краса та гігієна",
-                   "Офісні витрати", "Домашні тварини", "Подорожі", "Освіта та розвиток", "Перекази"]
-
-# list of full info about user_categories
-user_categories1 = {}
-
-# lst of the transfer transaction
-lst_transfer = {}
+lst_accounts = [str(number[0]) for number in numbers]
 
 
-# rewrite func dictionary of info about exp
-def rewrite_dict():
-    for i in user_categories:
-        user_categories1[i] = {}
-
-
+# ----------------------------------------------------------------------------------------------------------------
 class Menu:
 
     def __init__(self):
         self.main_menu_options = ["Управління категоріями витрат та доходів", "Управління рахунками",
-                                  "Управління витратами та доходами", "Пошук", "Вихід"]
+                                  "Управління витратами та доходами", "Вихід"]
         self.cat1 = CategoryOne
         self.cat2 = CategoryTwo
         self.cat3 = CategoryThree
-        self.cat4 = CategoryFour
 
-    # функціональний цикл меню, відповідає за роботу та виклик функцій обраних користувачем
+    # functional cycle of the menu, responsible for the operation and calling functions selected by the user
 
     @staticmethod
     def menu_loop(name_var, end, dictionary, menu_name=None):
@@ -80,21 +108,21 @@ class Menu:
 
     @staticmethod
     def the_end():
+        db.close()
         quit()
 
     def main_menu(self):
-        print("1.{} \n2.{} \n3.{} \n4.{} \n5.{} \nОберіть потрібну цифру: от 1 до 5: "
+        print("1.{} \n2.{} \n3.{} \n4.{} \nОберіть потрібну цифру: от 1 до 4: "
               .format(*self.main_menu_options))
         menu_dict = {
             1: self.cat1().menu_cat1,
             2: self.cat2().menu_cat2,
             3: self.cat3().menu_cat3,
-            4: self.cat4().menu_cat4,
-            5: self.the_end
+            4: self.the_end
         }
         try:
             choice = int(input("Введіть потрібний пункт: "))
-            self.menu_loop(choice, 5, menu_dict, self.main_menu)
+            self.menu_loop(choice, 4, menu_dict, self.main_menu)
         except ValueError:
             print("\nВи ввели неправильне значення. Спробуйте ще раз.\n")
             self.main_menu()
@@ -110,30 +138,20 @@ class Categories(ABC):
         self.menu_categories = ["Додати категорію", "Видалити категорію", "Змінити дані про категорію",
                                 "Перегляд списку категорій", "Назад\n"]
         self.user_categories = user_categories
-        self.user_categories1 = user_categories1
 
         # CategoryTwo
         self.bank_account = ["Створити новий рахунок", "Видалити рахунок", "Змінити дані рахунку",
                              "Переглянути список рахунків", "Назад\n"]
-        self.user_accounts = user_accounts
 
         # CategoryThree
-        self.income_expense_management = ["Додавання/видалення витрат до певного рахунку",
-                                          "Додавання/видалення прибутків до певного рахунку",
+        self.income_expense_management = ["Додати транзакцію",
+                                          "Видалити транзакцію",
                                           "Переведення грошей з рахунку на рахунок",
                                           "Перевірка витрат/прибутків за певний період",
                                           "Отримання статистики прибутків/витрат за певний період по днях та категоріях",
                                           "Назад\n"]
-        self.lst_transfer = lst_transfer
 
-        # 4 CategoryFour
-        self.search_transactions_op = ["Можливість пошуку категорій, витрат прибутків за категорією",
-                                       "Можливість пошуку прибутку витрати за сумою датою", "Назад\n"]
-
-        # universal
-        self.lst_accounts = lst_accounts
-
-    # відображення пунктів меню обраної категорії
+    # displaying menu items of the selected category
     @staticmethod
     def print_subcategory_menu(lst):
         for i, elem in enumerate(lst, start=1):
@@ -142,7 +160,7 @@ class Categories(ABC):
     def return_to_menu(self):
         self.menu.main_menu()
 
-    # перевіряє на тип даних та відправляє у виконання функції menu_loop
+    # checks for data type and sends to the menu_loop function
     def menu_universal(self, num, functional, menu_name):
         try:
             choice = int(input("Оберіть потрібний пункт: "))
@@ -150,15 +168,16 @@ class Categories(ABC):
         except ValueError:
             print("\nВи ввели неправильне значення. Спробуйте ще раз.\n")
 
-    # перевірка правильності номера рахунку
+    # checking the correctness of the account number
     def validate_account_num(self, num):
-        if num in self.lst_accounts:
+        if num in lst_accounts:
             return True
         else:
             print("Рахунок не знайдено")
             return False
 
-    # передаешь аргументов вторым тру если надо проверка на 3, если надо проверка на 2 тогда ничего не передаешь
+    # You pass the second argument as True if you need to check for 3.
+    # If you need to check for 2, then you don't pass anything.
     @staticmethod
     def validate_menu_choice(var, include_three=False):
         if not include_three:
@@ -173,37 +192,50 @@ class Categories(ABC):
                 return False
 
     def display_balance(self, account_num):
-        balance_info = self.user_accounts.get(account_num)
-        print("Баланс: {:,.2f} грн".format(balance_info.get("balance")))
+        cursor.execute("SELECT Balance FROM User_accounts WHERE Number = ?", (account_num,))
+        balance_info = cursor.fetchone()
+        print("Баланс: {:,.2f} грн".format(balance_info[0]))
 
     def display_account_info(self, account_num):
-        print("Номер Рахунку: {}".format(account_num))
-        print(f"ПІБ: {self.user_accounts[account_num]['name']}")
-        print(f"Тип: {self.user_accounts[account_num]['type']}")
+        cursor.execute("SELECT * FROM User_accounts WHERE Number = ?", (account_num,))
+        row = cursor.fetchone()
+        print(f"Номер Рахунку: {row[0]}")
+        print(f"Тип: {row[1]}")
+        print(f"ПІБ: {row[2]}")
         self.display_balance(account_num)
         print()
 
     def display_num_balance(self):
-        for key, val in self.user_accounts.items():
-            print("Номер рахунку: {}".format(key))
-            self.display_balance(key)
+        cursor.execute("SELECT NUMBER, Name FROM User_Accounts")
+        numbers_balance = cursor.fetchall()
+        for elem in numbers_balance:
+            print("Номер рахунку: {}".format(elem[0]))
+            print("ПІБ: {}".format(elem[1]))
+            self.display_balance(elem[0])
             print()
 
     def display_all_num(self):
-        for key, val in self.user_accounts.items():
-            print("Номер рахунку: {}".format(key))
+        for elem in lst_accounts:
+            cursor.execute("SELECT Name FROM User_Accounts WHERE Number = ?", (elem,))
+            name = cursor.fetchall()
+            print("Номер рахунку: {}".format(elem))
+            print("ПІБ: {}\n".format(name[0][0]))
 
     def display_transactions(self, account_num):
         """Отображение транзакций"""
-        transactions = self.user_accounts[account_num]["transactions"]
-        if len(transactions) != 0:
+        cursor.execute("SELECT * FROM TransactionAll WHERE Number = ?", (account_num,))
+        transaction_list = cursor.fetchall()
+        if len(transaction_list) != 0:
             self.display_account_info(account_num)
             print("Транзакції: ")
-            for i, transaction in enumerate(transactions):
+            count = 0
+            for elem in transaction_list:
+                count += 1
                 print(
-                    f"{i + 1}. {transaction['date']} | {transaction['category']} | {transaction['transaction']}"
-                    f" | id:{transaction['transaction_id']}")
-            self.visual()
+                    f"{count}. {elem[3]} | {elem[2]} | {elem[1]} | {elem[5]}"
+                    f" | id:{elem[4]}")
+                self.visual()
+            return True
         else:
             print("Транзакцій на рахунку: {} не знайдено\n".format(account_num))
             return False
@@ -234,7 +266,7 @@ class CategoryOne(Categories):
 
     def validate_name_categories(self, name):
         if name in self.user_categories:
-            print("Категорія с такою назвою вже існує.")
+            print("Категорія з назвою {} вже існує.".format(name))
             return False
         else:
             return True
@@ -245,28 +277,39 @@ class CategoryOne(Categories):
         name = input("Введіть назву нової категорії: ")
         if self.validate_name_categories(name):
             self.user_categories.append(name)
-            self.user_categories1[name] = {}
+            cursor.execute("INSERT INTO Category (Name) VALUES (?)", (name,))
+            db.commit()
             if name in self.user_categories:
-                print("Категорія додана")
+                print(f"Категорія {name} додана")
 
     def remove_category(self):
+        self.display_lst_user_cat()
         name = input("Введіть назву категорії: ")
         if name in self.user_categories:
             self.user_categories.remove(name)
+            cursor.execute("DELETE FROM Category WHERE Name = ?", (name,))
+            db.commit()
             if name not in self.user_categories:
-                print("Категорія видалена")
+                print(f"Категорія {name} видалена")
         elif name not in self.user_categories:
-            print("Категорію не знайдено")
+            print(f"Категорію з назвою {name} не знайдено")
 
     def update_category(self):
+        self.display_lst_user_cat()
         name = input("Введіть назву категорії, яку потрібно змінити: ")
         if name in self.user_categories:
+            print(f"Категорія з назвою {name} знайдена.")
             index = self.user_categories.index(name)
-            new_name = input("Введіть нову назву категорії: ")
-            self.user_categories[index] = new_name
-            print("Назван категорії оновлена")
+            while True:
+                new_name = input("Введіть нову назву: ")
+                if self.validate_name_categories(new_name):
+                    cursor.execute("UPDATE Category SET Name = ? WHERE Name = ?", (new_name, name))
+                    db.commit()
+                    self.user_categories[index] = new_name
+                    print(f"Назва категорії {name} змінена на {new_name}")
+                    break
         else:
-            print("Категорія не знайдена")
+            print(f"Категорія з назвою {name} не знайдена")
 
     def list_category(self):
         print("Список категорій:")
@@ -295,11 +338,11 @@ class CategoryTwo(Categories):
         return account_number
 
     def validate_new_account_num(self, account_num):
-        if account_num not in self.lst_accounts:
-            self.lst_accounts.append(account_num)
+        if account_num not in lst_accounts:
+            lst_accounts.append(account_num)
         else:
             account_num = self.generate_account_number()
-            self.lst_accounts.append(account_num)
+            lst_accounts.append(account_num)
         return account_num
 
     @staticmethod
@@ -321,37 +364,38 @@ class CategoryTwo(Categories):
 
     def input_type(self):
         while True:
-            account_type = input("Оберіть тип: \n1.дебетовий \n2.кредитний \nВведіть цифру 1 або 2: ")
+            account_type = input("Оберіть тип: \n1.Дебетовий \n2.Кредитний \nВведіть цифру 1 або 2: ")
             if self.validate_menu_choice(account_type):
-                return account_type
+                if account_type == "1":
+                    return "Дебетовий"
+                elif account_type == "2":
+                    return "Кредитний"
 
     def add_user_acc(self):
         account_num = self.generate_account_number()
         account_num = self.validate_new_account_num(account_num)
         account_name = self.input_name()
         account_type = self.input_type()
+        if account_num in lst_accounts:
+            cursor.execute("INSERT INTO User_Accounts (Number, Type, Name, Balance) "
+                           "VALUES (?, ?, ?, 0)", (account_num, account_type, account_name))
+            db.commit()
 
-        if account_type == "1":
-            self.user_accounts[account_num] = {"type": "Дебетовий", "name": account_name, "transactions": [],
-                                               "balance": 0.0}
-            print('Рахунок створено \n')
-            self.display_account_info(account_num)
-            print()
-
-        elif account_type == "2":
-            self.user_accounts[account_num] = {"type": "Кредитний", "name": account_name, "transactions": [],
-                                               "balance": 0.0}
-            print('Рахунок створено\n')
-            self.display_account_info(account_num)
-            print()
+        print('Рахунок створено\n')
+        self.display_account_info(account_num)
+        self.visual()
 
     def remove_user_acc(self):
         self.lst_user_acc()
         num_acc = input("Введіть номер рахунку для видалення: ")
-        if num_acc in self.user_accounts:
-            print("Рахунок видалено \n")
-            del self.user_accounts[num_acc]
-            self.lst_accounts.remove(num_acc)
+        if num_acc in lst_accounts:
+            cursor.execute("DELETE FROM User_Accounts WHERE Number = ?", (num_acc,))
+            lst_accounts.remove(num_acc)
+            if num_acc not in lst_accounts:
+                print(f"Рахунок {num_acc} видалено \n")
+                cursor.execute("DELETE FROM TransactionAll WHERE Number = ?", (num_acc,))
+                db.commit()
+
         else:
             print("Рахунок не знайдено \n")
             self.remove_user_acc()
@@ -364,13 +408,13 @@ class CategoryTwo(Categories):
                 # проверять ввод через функцию  validate_menu_choice
                 what_type = input("Оберіть пункт: ")
                 if what_type == "1":
-                    self.user_accounts[input_acc]["type"] = "Дебетовий"
+                    cursor.execute("UPDATE User_Accounts SET Type = ? WHERE Number = ?", ("Дебетовий", input_acc))
                     print("Тип рахунку змінено на Дебетовий \n")
                     self.display_account_info(input_acc)
                     return
 
                 elif what_type == "2":
-                    self.user_accounts[input_acc]["type"] = "Кредитний"
+                    cursor.execute("UPDATE User_Accounts SET Type = ? WHERE Number = ?", ("Кредитний", input_acc))
                     print("Тип рахунку змінено на Кредитний \n")
                     self.display_account_info(input_acc)
                     return
@@ -386,7 +430,7 @@ class CategoryTwo(Categories):
             while True:
                 new_name = input("Введіть новий ПІБ: ")
                 if self.validate_name(new_name):
-                    self.user_accounts[input_acc]["name"] = new_name
+                    cursor.execute("UPDATE User_Accounts SET Name = ? WHERE Number = ?", (new_name, input_acc))
                     print("Інформацію оновлено")
                     self.display_account_info(input_acc)
                     self.update_menu(input_acc)
@@ -418,19 +462,20 @@ class CategoryTwo(Categories):
     def update_user_acc(self):
         self.lst_user_acc()
         input_acc = input("Введіть номер рахунку: ")
-        if input_acc in self.user_accounts:
-            print("Рахунок знайдено\n")
+        if input_acc in lst_accounts:
+            print("Рахунок знайдено")
             self.display_account_info(input_acc)
             print()
             self.update_menu(input_acc)
+            db.commit()
         else:
             print("Рахунок не знайдено, спробуйте ще раз")
             self.update_user_acc()
 
     def lst_user_acc(self):
-        if len(self.user_accounts) == 0:
+        if len(lst_accounts) == 0:
             print("Рахунків нема")
-        for i in self.user_accounts:
+        for i in lst_accounts:
             self.display_account_info(i)
 
     def menu_cat2(self):
@@ -451,14 +496,11 @@ class CategoryThree(CategoryOne, CategoryTwo, Categories):
     @staticmethod
     def generate_random_date():
         """Генерация даты"""
-        current_time = datetime.datetime.now()
+        current_time = datetime.now()
         time_speed = random.uniform(0, 5)  # случайное значение скорости времени
-        step = datetime.timedelta(days=1)
+        step = timedelta(days=1)
         current_time += step * time_speed
-        year = current_time.year
-        month = current_time.month
-        day = current_time.day
-        return f"{day:02d}.{month:02d}.{year}"
+        return current_time.date()
 
     @staticmethod
     def generate_transaction_id():
@@ -477,74 +519,47 @@ class CategoryThree(CategoryOne, CategoryTwo, Categories):
                 break
             print("Категорія не знайдена, Спробуйте ще раз.")
 
-        pattern = r"^[+-]\d+$"
+        pattern = r"^[+\-]\d{1,10}$"
         while True:
-            trans = input("Введіть значення у форматі +/-сума, наприклад, "
-                          "+100 для додавання 100 грн або -100 для віднімання 100 грн: ")
-            if re.match(pattern, trans):
-                break
-            print("Введено неправильно. \nПриклад: +100, -100")
+            amount = input("Введіть значення транзакції у наступному форматі: "
+                           "+/- сума', наприклад, '+100' або '-100'."
+                           "Максимальна довжина - 10 символів: ")
 
-        return date, category_input, trans
+            trans = amount
+            if re.match(pattern, amount):
+                trans_type = "Дохід" if amount[0] == "+" else "Витрата"
+                return date, category_input, int(amount[1:]), trans_type, trans
+
+            print("Введено неправильно. \nПриклад: +100, -100 \nМаксимальна довжина - 10 символів ")
 
     @staticmethod
     def validate_date_input():
-        pattern = r"\d{2}\.\d{2}\.\d{4}"
+        pattern = r"\d{4}-\d{2}-\d{2}"  # Формат YYYY-MM-DD
         while True:
-            date_start_input = input("Введіть дату початку періоду: ")
+            date_start_input = input("Введіть дату початку періоду (в форматі YYYY-MM-DD): ")
             if re.match(pattern, date_start_input):
-                date_start = datetime.datetime.strptime(date_start_input, "%d.%m.%Y").date()
+                date_start = datetime.strptime(date_start_input, "%Y-%m-%d").date()
                 while True:
-                    date_end_input = input("Введіть дату кінця періоду: ")
+                    date_end_input = input("Введіть дату кінця початку періоду (в форматі YYYY-MM-DD): ")
                     if re.match(pattern, date_end_input):
-                        date_end = datetime.datetime.strptime(date_end_input, "%d.%m.%Y").date()
+                        date_end = datetime.strptime(date_end_input, "%Y-%m-%d").date()
                         if date_end > date_start:
                             return date_start, date_end
                         else:
                             print("Дата кінця періоду повинна бути більше дати початку періоду.")
+                    else:
+                        print("Введена неправильна дата")
             else:
                 print("Введена неправильна дата")
 
-    def delete_transaction(self, account_num, transaction, transactions):
-        transaction_amount = float(transaction["transaction"])
-        if transaction_amount > 0:
-            balance = self.user_accounts[account_num]["balance"] - transaction_amount
-        else:
-            balance = self.user_accounts[account_num]["balance"] + abs(transaction_amount)
-        transactions.remove(transaction)
-        self.user_accounts[account_num]["transactions"] = transactions
-        self.user_accounts[account_num]["balance"] = balance
-
-    def delete_transfer_transaction(self, account_num, transaction_id):
-        transaction_info = self.lst_transfer[transaction_id]
-        account_num1 = transaction_info["to_account"]  # если удаляется транзакция со счета получателя True
-        account_num2 = transaction_info["from_account"]  # если удаляется транзакция со счета отправителя False
-
-        verify = "2"
-        if account_num == account_num2:
-            verify = "1"
-
-        def delete(account_number):
-            transactions = self.user_accounts[account_number]["transactions"]
-            for elem in transactions:
-                if transaction_id == elem["transaction_id"]:
-                    print("id транзакции на втором счету найден")
-                    transaction_amount = float(elem["transaction"])
-                    if transaction_amount > 0:
-                        balance = self.user_accounts[account_number]["balance"] - transaction_amount
-                    else:
-                        balance = self.user_accounts[account_number]["balance"] + abs(transaction_amount)
-                    transactions.remove(elem)
-                    self.user_accounts[account_number]["transactions"] = transactions
-                    self.user_accounts[account_number]["balance"] = balance
-                    break
-
-        if verify == "1":
-            delete(account_num1)
-        elif verify == "2":
-            delete(account_num2)
-
-        del lst_transfer[transaction_id]
+    def input_id(self):
+        pattern = r"^TRX\d{6}$"
+        while True:
+            transaction_id = input("Введіть id транзакції: ")
+            if re.match(pattern, transaction_id):
+                return transaction_id
+            else:
+                print("Невірний формат id транзакції. Спробуйте ще раз.")
 
     # adding  transactions
     def add_transaction(self):
@@ -555,46 +570,38 @@ class CategoryThree(CategoryOne, CategoryTwo, Categories):
         account_num = self.input_num()
         self.display_balance(account_num)
         # Get the date, category, and transaction details from the user input
-        date, category, transaction = self.validate_money_input()
-        # Get the list of transactions and current balance for the selected account
-        transactions = self.user_accounts[account_num]["transactions"]
-        balance = self.user_accounts[account_num]["balance"]
-        # Add the new transaction to the list of transactions and update the balance
-        transactions.append(
-            {"transaction_id": trans_id, "date": date, "category": category, "transaction": transaction})
-        balance += float(transaction)
-        # Update the transactions list and balance for the selected account
-        self.user_accounts[account_num]["transactions"] = transactions
-        self.user_accounts[account_num]["balance"] = balance
-        print("Транзакція додана: {} | {} | {}\n".format(date, category, transaction))
+        date, category, amount, trans_type, trans = self.validate_money_input()
+
+        cursor.execute("INSERT INTO TransactionAll (Number, Type, Category, TransactionDate, TransactionID, Amount) "
+                       "VALUES (?,?,?,?,?,?)", (account_num, trans_type, category, date, trans_id, amount))
+        if trans_type == "Витрата":
+            cursor.execute("UPDATE User_Accounts SET Balance = Balance - ? WHERE Number = ?", (amount, account_num))
+        elif trans_type == "Дохід":
+            cursor.execute("UPDATE User_Accounts SET Balance = Balance + ? WHERE Number = ?", (amount, account_num))
+
+        print("Транзакція додана: {} | {} | {}\n".format(date, category, trans))
+        db.commit()
         self.display_account_info(account_num)
-        # Add the new transaction to the user_categories1 dictionary
-        if category not in self.user_categories1:
-            self.user_categories1[category] = {}
-        if account_num not in self.user_categories1[category]:
-            self.user_categories1[category][account_num] = {}
-        if date not in self.user_categories1[category][account_num]:
-            self.user_categories1[category][account_num][date] = []
-        new_transaction = {"transaction": transaction, "transaction_id": trans_id}
-        self.user_categories1[category][account_num][date].append(new_transaction)
 
     # delete transactions
     def delete_transactions(self):
         self.lst_user_acc()
         account_num = self.input_num()
-        transactions = user_accounts[account_num]["transactions"]
         if self.display_transactions(account_num):
-            transaction_id = input("Введіть id транзакції: ")
-            for elem in transactions:
-                if transaction_id == elem["transaction_id"]:
-                    if elem["category"] == "Перекази":
-                        transfer_transaction = elem
-                        self.delete_transaction(account_num, transfer_transaction, transactions)
-                        self.delete_transfer_transaction(account_num, transaction_id)
-                    else:
-                        transaction = elem
-                        self.delete_transaction(account_num, transaction, transactions)
-            print("Транзакція видалена")
+            transaction_id = self.input_id()
+            cursor.execute("SELECT * FROM TransactionAll WHERE TransactionID = ?", (transaction_id,))
+            list_transaction_del = cursor.fetchall()
+            for elem in list_transaction_del:
+                if elem[1] == "Витрата":
+                    cursor.execute("UPDATE User_Accounts SET Balance = Balance + ? WHERE Number = ? ",
+                                   (elem[5], elem[0]))
+                if elem[1] == "Дохід":
+                    cursor.execute("UPDATE User_Accounts SET Balance = Balance - ? WHERE Number = ? ",
+                                   (elem[5], elem[0]))
+
+            cursor.execute("DELETE FROM TransactionAll WHERE TransactionID = ?", (transaction_id,))
+            db.commit()
+            print("Транзакція видалена \n")
 
     # transfer money between accounts
     def transfer_money(self):
@@ -609,7 +616,9 @@ class CategoryThree(CategoryOne, CategoryTwo, Categories):
         transfer_id = self.generate_transaction_id()
         date1 = self.generate_random_date()
         category = "Перекази"
-        if user_accounts[from_num1]["balance"] != 0.0:
+        cursor.execute("SELECT Balance FROM User_Accounts WHERE Number = {}".format(from_num1))
+        balance_num1 = cursor.fetchall()
+        if int(balance_num1[0][0]) != 0:
             while True:
                 amount = float(input("Введіть суму для переводу: "))
                 if amount > 0:
@@ -617,39 +626,36 @@ class CategoryThree(CategoryOne, CategoryTwo, Categories):
                 else:
                     print("Сума повинна бути більше нуля. Будь ласка, введіть коректну суму.")
 
-            # format amount transaction
-            trans1 = "-{:.2f}".format(amount)
-            trans2 = "+{:.2f}".format(amount)
+            # add the new transaction to the db in Table | TransactionAll | Transaction_Transfer
+            if int(balance_num1[0][0]) > amount:
+                # Начать транзакцию
+                cursor.execute('BEGIN TRANSACTION')
 
-            # add the new transaction to the user_categories and lst_transfer
-            if user_accounts[from_num1]["balance"] >= amount:
-                user_accounts[from_num1]["balance"] -= amount
-                user_accounts[to_num2]["balance"] += amount
-                user_accounts[from_num1]["transactions"].append(
-                    {"transaction_id": transfer_id, "date": date1, "category": category, "transaction": trans1})
+                # INSERT DATA IN TransactionAll FROM_NUM
+                cursor.execute(
+                    'INSERT INTO TransactionAll (Number, Type, Category, TransactionDate, TransactionID, Amount) VALUES'
+                    '(?, ?, ?, ?, ?, ?)',
+                    (from_num1, "Витрата", category, date1, transfer_id, amount))
+                # INSERT DATA IN TransactionAll TO_NUM
+                cursor.execute(
+                    'INSERT INTO TransactionAll (Number, Type, Category, TransactionDate, TransactionID, Amount) VALUES'
+                    '(?, ?, ?, ?, ?, ?)',
+                    (to_num2, "Дохід", category, date1, transfer_id, amount))
+                # UPDATE BALANCE FROM_NUM
+                cursor.execute(
+                    "UPDATE User_Accounts SET Balance = Balance - {} WHERE Number = {}".format(amount, from_num1))
+                # UPDATE BALANCE TO_NUM
+                cursor.execute(
+                    "UPDATE User_Accounts SET Balance = Balance + {} WHERE Number = {}".format(amount, to_num2))
 
-                user_accounts[to_num2]["transactions"].append(
-                    {"transaction_id": transfer_id, "date": date1, "category": category, "transaction": trans2})
-                print("\nПереказ виконано\n")
+                # INSERT DATA IN Transaction_Transfer
+                cursor.execute(
+                    'INSERT INTO Transaction_Transfer (From_number, To_number, TransactionDate, TransactionID, Amount ) VALUES (?, ?, ?, ?, ?)',
+                    (from_num1, to_num2, date1, transfer_id, amount))
 
-                lst_transfer[transfer_id] = {"date": date1, "from_account": from_num1, "to_account": to_num2,
-                                             "amount": amount}
-                self.visual()
-                # Add the new transactions to the user_categories1 dictionary
-                if from_num1 not in self.user_categories1[category]:
-                    self.user_categories1[category][from_num1] = {}
-                if date1 not in self.user_categories1[category][from_num1]:
-                    self.user_categories1[category][from_num1][date1] = []
-                if to_num2 not in self.user_categories1[category][from_num1][date1]:
-                    self.user_categories1[category][from_num1][date1].append(
-                        {"transaction_id": transfer_id, "transaction": trans2})
-                if to_num2 not in self.user_categories1[category]:
-                    self.user_categories1[category][to_num2] = {}
-                if date1 not in self.user_categories1[category][to_num2]:
-                    self.user_categories1[category][to_num2][date1] = []
-                if from_num1 not in self.user_categories1[category][to_num2][date1]:
-                    self.user_categories1[category][to_num2][date1].append(
-                        {"transaction_id": transfer_id, "transaction": trans1})
+                # Закончить транзакцию
+                cursor.execute('COMMIT')
+                print("Транзакція пройшла успішно")
 
             else:
                 print("Недостатньо коштів на рахунку")
@@ -657,45 +663,37 @@ class CategoryThree(CategoryOne, CategoryTwo, Categories):
             print("Баланс рахунку: {} пустий".format(from_num1))
 
     # info about  expense\income time interval
-    """ Добавить от дата конца не может быть больше даты начала"""
 
     def get_expenses_income_by_period(self):
         self.display_all_num()
         num = self.input_num()
         print("Для перевірки витрат і прибутків за певний період, будь ласка, введіть дату початку періоду та дату "
-              "його завершення. Формат дати має бути наступним: день.місяць.рік (наприклад, 01.05.2023)")
+              "його завершення. Формат дати має бути наступним: рік-місяць-день (наприклад, 2023-05-16)")
         start_date, end_date = self.validate_date_input()
         income = 0
         expense = 0
-        transactions = self.user_accounts[num]["transactions"]
-        for transaction in transactions:
-            transaction_date = datetime.datetime.strptime(transaction['date'], "%d.%m.%Y").date()
-            transaction_amount = float(transaction['transaction'][1:])
-            transaction_type = transaction['transaction'][0]
-            if start_date <= transaction_date <= end_date:
-                if transaction_type == '+':
-                    income += transaction_amount
-                elif transaction_type == '-':
-                    expense += transaction_amount
+        cursor.execute(
+            "SELECT * FROM TransactionAll WHERE Number = ? AND TransactionDate BETWEEN ? AND ?",
+            (num, start_date, end_date))
+        list_transaction = cursor.fetchall()
+        for elem in list_transaction:
+            if elem[1] == "Дохід":
+                income += elem[5]
+            else:
+                expense += elem[5]
         print(f"Для введеного Вами періоду часу, загальна сума витрат становить {expense:,.2f} гривень.")
         print(f"Також, загальний прибуток за цей період склав {income:,.2f} гривень. ")
         print("Дякую, що користуєтесь нашим сервісом!")
 
-    # {'Їжа': {'12345678': {'09.05.2023': [{'transaction': '+100', 'transaction_id': 'TRX153211'}]}}
     # info about all times expense\income in categories
     def get_statistics(self):
         self.display_all_num()
-        print()
         num = self.input_num()
-        for category, accounts in self.user_categories1.items():
-            if num in accounts:
-                for account_num, dates in accounts.items():
-                    for date, lst_transaction in dates.items():
-                        amount = 0
-                        for elem in lst_transaction:
-                            count = float(elem["transaction"])
-                            amount += count
-                            print("Категорія: {} | Дата: {} | Сумма: {}".format(category, date, amount))
+        cursor.execute("SELECT * FROM TransactionAll WHERE Number = ? ", (num,))
+        statistics_data = cursor.fetchall()
+        for elem in statistics_data:
+            print("Категорія: {} | Дата: {} | Тип: {} | Сумма: {}".format(elem[2], elem[3], elem[1], elem[5]))
+
         print()
 
     def menu_cat3(self):
@@ -712,27 +710,5 @@ class CategoryThree(CategoryOne, CategoryTwo, Categories):
             self.menu_universal(6, functional, self.menu_cat3)
 
 
-class CategoryFour(Categories):
-
-    def search_by_category(self):
-        print("welcome to the search_by_category \n Ніхуя НЕМА ТУТ БО ПІШОВ НАХУЙ")
-        pass
-
-    def search_by_amount_date(self):
-        print("welcome to the search_by_amount_date \n Ніхуя НЕМА ТУТ БО ПІШОВ НАХУЙ ")
-        pass
-
-    def menu_cat4(self):
-        functional = {
-            1: self.search_by_category,
-            2: self.search_by_amount_date,
-            3: self.return_to_menu
-        }
-        while True:
-            self.print_subcategory_menu(self.search_transactions_op)
-            self.menu_universal(3, functional, self.menu_cat4)
-
-
-rewrite_dict()
 test = Menu()
 test.main_menu()
